Refactor Tradfri LightGroup, part 1
